Move my_singleton to utils folder.
