+++ src/utils/my_singleton.cpp
